MB-60687: Make initial value type explicit for std::accumulate
As seen in MB-60625, passing `0` will make the accumulator type `int`,
which can result in overflow.

Reviewed-on: https://review.couchbase.org/c/platform/+/205028
Tested-by: Build Bot <build@couchbase.com>

+++ src/histogram.cc
@@ -28,7 +28,7 @@ void Histogram<T, Limits>::reset() {
 template <typename T, template <class> class Limits>
 size_t Histogram<T, Limits>::total() const {
     HistogramBinSampleAdder<T, Limits> a;
-    return std::accumulate(begin(), end(), 0, a);
+    return std::accumulate(begin(), end(), size_t(0), a);
 }
 
 template <typename T, template <class> class Limits>
